Basically finished the Teacher GUI
We still have to create the logic for these methods. I created some of them, but most of them are incomplete. I am counting on y'all to fix them because I still have to do the Student GUI and work on the rest of my tasks. Also, we need to update our model classes to suit the GUI properly. We'll discuss this tomorrow.